+++ CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 -   user_settings_set_changed_with_key and user_settings_set_changed_with_id functions.
 -   always_mark_changed parameter to user_settings_set_from_json function.
+-   The "has_changed_recently" flag is now stored persistently.
+-   Shell commands to list only changed settings and to clear the changed flag.
 
 ## [1.5.0] - 2023-04-11